Move config types & allow multiple processes to be run
